refactor(NODE-3717): reorganize tests pt5
